git-svn-id: http://svn.svnkit.com/repos/sqljet@1291 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ branches/1.1.x/sqljet/src/main/java/org/tmatesoft/sqljet/core/internal/btree/SqlJetBtree.java
@@ -1807,7 +1807,7 @@ public class SqlJetBtree implements ISqlJetBtree {
                 pToPage = pBtTo.pPager.getPage(i);
                 pToPage.write();
 
-                for (iOff = (i - 1) * nToPageSize; iOff < i * nToPageSize; iOff += nFromPageSize) {
+                for (iOff = (i - 1) * (long)nToPageSize; iOff < i * (long)nToPageSize; iOff += nFromPageSize) {
                     ISqlJetPage pFromPage = null;
                     int iFrom = (int) (iOff / nFromPageSize) + 1;
